backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.36
asyncpg==0.29.0
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
anthropic==0.40.0
langgraph==0.2.55
langchain-anthropic==0.3.0
celery==5.4.0
redis==5.1.1
pandas==2.2.3
python-dotenv==1.0.1
